refactor admin_definition to eliminate circular dependencies
